+++ clang/docs/ReleaseNotes.rst
@@ -1168,6 +1168,7 @@ OpenMP Support
 - Added support for 'omp assume' directive.
 - Added support for 'omp scope' directive.
 - Added support for allocator-modifier in 'allocate' clause.
+- Added support for 'omp stripe' directive.
 
 Improvements
 ^^^^^^^^^^^^